Async mail service to reduce email sending time
