+++ service-mail/src/main/java/sg/ncl/service/mail/domain/AsyncMailService.java
@@ -1,0 +1,12 @@
+package sg.ncl.service.mail.domain;
+
+import sg.ncl.service.mail.data.jpa.EmailEntity;
+
+/**
+ * Created by dcsjnh on 24/4/2017.
+ */
+public interface AsyncMailService {
+
+    void send(EmailEntity emailEntity);
+
+}
